requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
python-dotenv

